[FIX] Нельзя удалить отдел с сотрудниками

+++ src/main/java/project/controller/SubdepartmentController.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.*;
 import project.model.Subdepartment;
+import project.service.StudentServiceImpl;
 import project.service.SubdepartmentServiceImpl;
 
 import java.util.List;
@@ -16,6 +17,7 @@ import static project.exceptions.Validator.validateDepartmentId;
 public class SubdepartmentController {
 
     private final SubdepartmentServiceImpl service;
+    private final StudentServiceImpl studentService;
 
     @GetMapping("/dep_{N}/subdep/data")
     public List<Subdepartment> getAll(@PathVariable("N") String depId,
@@ -46,6 +48,9 @@ public class SubdepartmentController {
                                         @PathVariable("id") short id,
                                         @RequestHeader(value = HttpHeaders.AUTHORIZATION, defaultValue = "") String jwtToken) {
         Integer departmentId = validateDepartmentId(depId);
+        if (studentService.getStudentExistsWithSubdepartmentId(departmentId, id)) {
+            throw new IllegalArgumentException("Существует работник из этого отдела. Необходимо добавить его в другой отдел.");
+        }
         service.deleteById(departmentId, id);
     }
